+++ android/src/main/java/com/reactnativestripesdk/EmbeddedPaymentElementViewManager.kt
@@ -24,7 +24,6 @@ import com.reactnativestripesdk.utils.mapToPreferredNetworks
 import com.reactnativestripesdk.utils.parseCustomPaymentMethods
 import com.stripe.android.ExperimentalAllowsRemovalOfLastSavedPaymentMethodApi
 import com.stripe.android.paymentelement.EmbeddedPaymentElement
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.paymentsheet.CardFundingFilteringPrivatePreview
 import com.stripe.android.paymentsheet.PaymentSheet
 import org.json.JSONArray
@@ -99,7 +98,6 @@ class EmbeddedPaymentElementViewManager :
   @SuppressLint("RestrictedApi")
   @OptIn(
     ExperimentalAllowsRemovalOfLastSavedPaymentMethodApi::class,
-    ExperimentalCustomPaymentMethodsApi::class,
     CardFundingFilteringPrivatePreview::class,
   )
   private fun parseElementConfiguration(