+++ android/src/main/java/com/reactnativestripesdk/PaymentSheetManager.kt
@@ -45,7 +45,6 @@ import com.stripe.android.paymentelement.ConfirmCustomPaymentMethodCallback
 import com.stripe.android.paymentelement.CreateIntentWithConfirmationTokenCallback
 import com.stripe.android.paymentelement.CustomPaymentMethodResult
 import com.stripe.android.paymentelement.CustomPaymentMethodResultHandler
-import com.stripe.android.paymentelement.ExperimentalCustomPaymentMethodsApi
 import com.stripe.android.paymentelement.PaymentMethodOptionsSetupFutureUsagePreview
 import com.stripe.android.paymentsheet.CardFundingFilteringPrivatePreview
 import com.stripe.android.paymentsheet.CreateIntentCallback
@@ -68,7 +67,6 @@ import kotlin.coroutines.resume
 @OptIn(
   ReactNativeSdkInternal::class,
   ExperimentalAllowsRemovalOfLastSavedPaymentMethodApi::class,
-  ExperimentalCustomPaymentMethodsApi::class,
   CardFundingFilteringPrivatePreview::class,
 )
 class PaymentSheetManager(
@@ -90,7 +88,6 @@ class PaymentSheetManager(
   private var keepJsAwake: KeepJsAwakeTask? = null
 
   @SuppressLint("RestrictedApi")
-  @OptIn(ExperimentalCustomPaymentMethodsApi::class)
   override fun onCreate() {
     val activity = getCurrentActivityOrResolveWithError(initPromise) ?: return
     val merchantDisplayName = arguments.getString("merchantDisplayName").orEmpty()
@@ -502,7 +499,6 @@ class PaymentSheetManager(
     } ?: run { resolvePresentPromise(map) }
   }
 
-  @OptIn(ExperimentalCustomPaymentMethodsApi::class)
   override fun onConfirmCustomPaymentMethod(
     customPaymentMethod: PaymentSheet.CustomPaymentMethod,
     billingDetails: PaymentMethod.BillingDetails,