+++ example-stripe-connect/index.ts
@@ -1,1 +1,4 @@
 import 'expo-router/entry';
+import { LogBox } from 'react-native';
+
+LogBox.ignoreAllLogs();